Deleted GetUsers,SaveUsers and made JsonHandlerService
